requirements.txt
importlib-metadata<5.0
pytest
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1


matplotlib
setuptools

hydra-core
hydra_colorlog

seaborn
torch>=2.2.0
normflows
torchvision
pyro-ppl
pytorch-lightning
mlflow
numpy
scikit-learn
torchdyn

omegaconf
tqdm
ray[tune]==1.13.0
POT
git+https://github.com/fadel/pytorch_ema